requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.13.0=pypi_0
altair=4.1.0=pypi_0
argon2-cffi=20.1.0=py38h2bbff1b_1
astor=0.8.1=pypi_0
astunparse=1.6.3=pypi_0
async_generator=1.10=pyhd3eb1b0_0
attrs=21.2.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
base58=2.1.0=pypi_0
bleach=3.3.0=pyhd3eb1b0_0
blinker=1.4=pypi_0
ca-certificates=2021.5.25=haa95532_1
cachetools=4.2.2=pypi_0
certifi=2021.5.30=py38haa95532_0
cffi=1.14.5=py38hcd4344a_0
chardet=4.0.0=pypi_0
click=7.1.2=pypi_0
colorama=0.4.4=pyhd3eb1b0_0
console_shortcut=0.1.1=4
debugpy=1.3.0=pypi_0
decorator=5.0.9=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.3=pypi_0
flatbuffers=1.12=pypi_0
gast=0.4.0=pypi_0
gitdb=4.0.7=pypi_0
gitpython=3.1.18=pypi_0
google-auth=1.32.1=pypi_0
google-auth-oauthlib=0.4.4=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.34.1=pypi_0
h5py=3.1.0=pypi_0
idna=2.10=pypi_0
importlib-metadata=3.10.0=py38haa95532_0
importlib_metadata=3.10.0=hd3eb1b0_0
ipykernel=6.0.1=pypi_0
ipython=7.25.0=pypi_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.3=pypi_0
jedi=0.18.0=pypi_0
jinja2=3.0.1=pyhd3eb1b0_0
jsonschema=3.2.0=py_2
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.7.1=py38haa95532_0
jupyterlab-widgets=1.0.0=pypi_0
jupyterlab_pygments=0.1.2=py_0
keras-nightly=2.5.0.dev2021032900=pypi_0
keras-preprocessing=1.1.2=pypi_0
libsodium=1.0.18=h62dcd97_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.3.4=pypi_0
markupsafe=2.0.1=py38h2bbff1b_0
matplotlib-inline=0.1.2=pypi_0
mistune=0.8.4=py38he774522_1000
msys2-conda-epoch=20160418=1
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.1.0=py38haa95532_0
nbformat=5.1.3=pyhd3eb1b0_0
nest-asyncio=1.5.1=pyhd3eb1b0_0
notebook=6.4.0=py38haa95532_0
numpy=1.19.5=pypi_0
oauthlib=3.1.1=pypi_0
openssl=1.1.1k=h2bbff1b_0
opt-einsum=3.3.0=pypi_0
packaging=20.9=pyhd3eb1b0_0
pandocfilters=1.4.3=py38haa95532_1
parso=0.8.2=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.3.0=pypi_0
pip=21.1.3=py38haa95532_0
prometheus_client=0.11.0=pyhd3eb1b0_0
prompt-toolkit=3.0.19=pypi_0
protobuf=3.17.3=pypi_0
pyarrow=4.0.1=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=py_2
pydeck=0.6.2=pypi_0
pygments=2.9.0=pyhd3eb1b0_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyrsistent=0.18.0=pypi_0
python=3.8.10=hdbf39b2_7
python-dateutil=2.8.1=pyhd3eb1b0_0
pytz=2021.1=pypi_0
pywin32=301=pypi_0
pywinpty=1.1.3=pypi_0
pyzmq=22.1.0=pypi_0
requests=2.25.1=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
send2trash=1.7.1=pypi_0
setuptools=52.0.0=py38haa95532_0
six=1.15.0=pypi_0
smmap=4.0.0=pypi_0
sqlite=3.36.0=h2bbff1b_0
streamlit=0.84.0=pypi_0
tensorboard=2.5.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorflow=2.5.0=pypi_0
tensorflow-estimator=2.5.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.10.1=pypi_0
testpath=0.5.0=pyhd3eb1b0_0
toml=0.10.2=pypi_0
toolz=0.11.1=pypi_0
tornado=6.1=py38h2bbff1b_0
traitlets=5.0.5=pyhd3eb1b0_0
typing-extensions=3.7.4.3=pypi_0
tzlocal=2.1=pypi_0
urllib3=1.26.6=pypi_0
validators=0.18.2=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=2.1.3=pypi_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=pypi_0
werkzeug=2.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=pypi_0
wincertstore=0.2=py38_0
winpty=0.4.3=4
wrapt=1.12.1=pypi_0
zeromq=4.3.3=ha925a31_3
zipp=3.4.1=pyhd3eb1b0_0
